Track marked nodes in BTree file.
It's pretty storage efficient to store the mark bits in the FlatFile.
But this increases the risk of data corruption and is slower since we
have very strict sync policies now on the FlatFile. The mark data is
very volatile and not too precious. We also might want to support
multiple parallel each() runs, so we would need a separate mark list
for each run.