Fix: Fix a storage space leak in the BTree files
There was a storage space leak in the BTree files. This is used for the
local database index and recently the free space list. This patch
removes the leak. Existing files can continue to be used. If you want to
reclaim the lost storage space you need to call Store.repair() once.