Fix: Eliminate a race between the GC and loads from files.
It seems that the GC is not completely removing all objects and calling
collect() on them atomically. This could lead to situations where the
object is no longer in memory but not yet written to files. If we now
try to load if from disk again, we might pick up old data.